refactor(cache): move cache functions to their own file
